+++ packages/language-server/test/unit/completion-items.test.ts
@@ -30,7 +30,7 @@ describe("the UI5 language assistant Code Completion Services", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);