+++ packages/language-server/test/unit/hover.test.ts
@@ -23,7 +23,7 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);