+++ packages/xml-views-completion/test/unit/api.test.ts
@@ -23,7 +23,7 @@ describe("The `getXMLViewCompletions()` api", () => {
   beforeAll(async function () {
     REAL_UI5_MODEL = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(REAL_UI5_MODEL);