+++ packages/xml-views-completion/test/unit/providers/attributeValue/enum.test.ts
@@ -14,7 +14,7 @@ describe("The ui5-language-assistant xml-views-completion", () => {
   beforeAll(async function () {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);