+++ packages/xml-views-completion/test/unit/providers/attributeValue/namespace.test.ts
@@ -31,7 +31,7 @@ describe("The ui5-editor-tools xml-views-completion", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);