+++ packages/xml-views-validation/test/unit/api.test.ts
@@ -17,7 +17,7 @@ describe("the ui5 xml views validations API", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);