+++ packages/xml-views-validation/test/unit/validators/attributes/unknown-attribute-key.test.ts
@@ -21,7 +21,7 @@ describe("the unknown attribute name validation", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);