+++ packages/xml-views-validation/test/unit/validators/element/cardinality-of-aggregation.test.ts
@@ -22,7 +22,7 @@ describe("the cardinality aggregation validation", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);