+++ packages/xml-views-validation/test/unit/validators/element/non-stable-id.test.ts
@@ -22,7 +22,7 @@ describe("the use of non stable id validation", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);