+++ packages/xml-views-validation/test/unit/validators/element/use-of-deprecated-class.test.ts
@@ -15,7 +15,7 @@ describe("the use of deprecated class validation", () => {
   beforeAll(async () => {
     ui5SemanticModel = await generateModel({
       framework: "SAPUI5",
-      version: "1.71.60",
+      version: "1.71.61",
       modelGenerator: generate,
     });
     appContext = getDefaultContext(ui5SemanticModel);